requirements.txt
streamlit>=1.31.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
requests>=2.31.0
uagents>=0.11.0
agentmail-sdk>=0.1.0
cloudflare>=2.19.0
plotly>=5.19.0
pydantic>=2.6.0
python-jose>=3.3.0
aiohttp>=3.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
